mv v2 to root
