+++ main.go
@@ -9,13 +9,12 @@ import (
 	"os"
 	"os/signal"
 	"strings"
-	"syscall"
-	"text/tabwriter"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/oklog/run"
+	"github.com/peterbourgon/usage"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
@@ -24,15 +23,15 @@ var version = "dev"
 func main() {
 	fs := flag.NewFlagSet("fastly-exporter", flag.ExitOnError)
 	var (
-		token      = fs.String("token", "", "Fastly API token")
+		token      = fs.String("token", "", "Fastly API token (required)")
 		serviceIDs = stringslice{}
 		addr       = fs.String("endpoint", "http://127.0.0.1:8080/metrics", "Prometheus /metrics endpoint")
-		namespace  = fs.String("namespace", "", "Prometheus namespace")
-		subsystem  = fs.String("subsystem", "", "Prometheus subsystem")
-		debug      = fs.Bool("debug", false, "log debug information")
+		namespace  = fs.String("namespace", "", "Prometheus namespace (optional)")
+		subsystem  = fs.String("subsystem", "", "Prometheus subsystem (optional)")
+		debug      = fs.Bool("debug", false, "Log debug information")
 	)
-	fs.Var(&serviceIDs, "service", "Fastly service ID (repeatable)")
-	fs.Usage = usageFor(fs, "fastly-exporter [flags]")
+	fs.Var(&serviceIDs, "service", "Specific Fastly service ID (optional, repeatable)")
+	fs.Usage = usage.For(fs, "fastly-exporter [flags]")
 	fs.Parse(os.Args[1:])
 
 	var logger log.Logger
@@ -49,15 +48,10 @@ func main() {
 		level.Error(logger).Log("err", "-token is required")
 		os.Exit(1)
 	}
-	if len(serviceIDs) <= 0 {
-		level.Error(logger).Log("err", "at least one -service ID is required")
-		os.Exit(1)
-	}
 
-	level.Debug(logger).Log("msg", "looking up service names")
-	serviceNames := getServiceNames(*token, serviceIDs, log.With(logger, "query", "api.fastly.com"))
-	for service, name := range serviceNames {
-		level.Info(logger).Log("fastly_service", service, "name", name)
+	var metrics prometheusMetrics
+	{
+		metrics.register(*namespace, *subsystem)
 	}
 
 	var promURL *url.URL
@@ -71,27 +65,74 @@ func main() {
 		level.Info(logger).Log("prometheus_addr", promURL.Host, "path", promURL.Path, "namespace", *namespace, "subsystem", *subsystem)
 	}
 
-	var m prometheusMetrics
+	var cache *nameCache
 	{
-		m.register(*namespace, *subsystem)
+		cache = newNameCache()
+	}
+
+	var manager *monitorManager
+	{
+		var (
+			postprocess = func() {}                          // only used for tests
+			rtClient    = &http.Client{Timeout: time.Minute} // rt.fastly.com blocks awhile by design
+		)
+		manager = newMonitorManager(rtClient, *token, cache, metrics, postprocess, log.With(logger, "component", "monitors"))
+	}
+
+	var apiClient *http.Client
+	{
+		apiClient = &http.Client{
+			Timeout: 10 * time.Second, // api.fastly.com should be fast
+		}
+	}
+
+	var queryer *serviceQueryer
+	{
+		queryer = newServiceQueryer(*token, serviceIDs, cache, manager)
+		if err := queryer.refresh(apiClient); err != nil { // first refresh must succeed
+			level.Error(logger).Log("during", "initial service refresh", "err", err)
+			os.Exit(1)
+		}
 	}
 
 	var g run.Group
 	{
-		for serviceID, serviceName := range serviceNames {
-			var (
-				ctx, cancel = context.WithCancel(context.Background())
-				serviceID   = serviceID   // shadow copy
-				serviceName = serviceName // shadow copy
-			)
-			g.Add(func() error {
-				return queryLoop(ctx, *token, serviceID, serviceName, &m, log.With(logger, "query", "rt.fastly.com", "service", serviceName))
-			}, func(error) {
-				cancel()
-			})
-		}
+		// Every minute, query Fastly for new services and their names.
+		// Update our name cache and managed monitors accordingly.
+		var (
+			ctx, cancel = context.WithCancel(context.Background())
+			ticker      = time.NewTicker(time.Minute)
+		)
+		g.Add(func() error {
+			for {
+				select {
+				case <-ctx.Done():
+					return ctx.Err()
+
+				case <-ticker.C:
+					if err := queryer.refresh(apiClient); err != nil {
+						level.Warn(logger).Log("during", "service refresh", "err", err)
+					}
+				}
+			}
+		}, func(error) {
+			ticker.Stop()
+			cancel()
+		})
 	}
 	{
+		// A pseudo-actor that exists only to tear down all managed monitors.
+		ctx, cancel := context.WithCancel(context.Background())
+		g.Add(func() error {
+			<-ctx.Done()
+			manager.stopAll()
+			return ctx.Err()
+		}, func(error) {
+			cancel()
+		})
+	}
+	{
+		// Serve Prometheus metrics over HTTP.
 		mux := http.NewServeMux()
 		mux.Handle(promURL.Path, promhttp.Handler())
 		server := http.Server{
@@ -107,43 +148,24 @@ func main() {
 		})
 	}
 	{
-		ctx, cancel := context.WithCancel(context.Background())
+		// Catch ctrl-C.
+		var (
+			ctx, cancel = context.WithCancel(context.Background())
+			c           = make(chan os.Signal, 1)
+		)
+		signal.Notify(c, os.Interrupt)
 		g.Add(func() error {
-			c := make(chan os.Signal, 1)
-			signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
 			select {
-			case sig := <-c:
-				return fmt.Errorf("received signal %s", sig)
 			case <-ctx.Done():
 				return ctx.Err()
+			case sig := <-c:
+				return fmt.Errorf("received signal %s", sig)
 			}
 		}, func(error) {
 			cancel()
 		})
 	}
 	level.Info(logger).Log("exit", g.Run())
-}
-
-func usageFor(fs *flag.FlagSet, short string) func() {
-	return func() {
-		fmt.Fprintf(os.Stderr, "USAGE\n")
-		fmt.Fprintf(os.Stderr, "  %s\n", short)
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "FLAGS\n")
-		w := tabwriter.NewWriter(os.Stderr, 0, 2, 2, ' ', 0)
-		fs.VisitAll(func(f *flag.Flag) {
-			def := f.DefValue
-			if def == "" {
-				def = "..."
-			}
-			fmt.Fprintf(w, "\t-%s %s\t%s\n", f.Name, def, f.Usage)
-		})
-		w.Flush()
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "VERSION\n")
-		fmt.Fprintf(os.Stderr, "  %s\n", version)
-		fmt.Fprintf(os.Stderr, "\n")
-	}
 }
 
 type stringslice []string
@@ -158,4 +180,8 @@ func (ss *stringslice) String() string {
 		return "..."
 	}
 	return strings.Join(*ss, ", ")
+}
+
+type httpClient interface {
+	Do(*http.Request) (*http.Response, error)
 }
